feat(community): added phase select option on member vizualiation table + move table to community
