+++ src/models/user_startup.ts
@@ -1,0 +1,4 @@
+interface UserStartup {
+    user_id: string,
+    startup_id: string
+}
